Started to change cards for github

+++ backend/hooks/routes.py
@@ -2,7 +2,7 @@ from flask import Blueprint, request
 from flask_restful import Resource
 from backend import api
 from backend.config import API
-from backend.hooks.utils import edit_test_json, get_files, md_to_json, parse_module
+from backend.hooks.utils import edit_test_json, get_files, md_to_json, parse_activity, parse_module
 import requests
 
 # Blueprint for hooks
@@ -11,6 +11,10 @@ hooks_bp = Blueprint("hooks", __name__)
 # git commit -m "Testing hooks"
 # git push origin test
 
+# git add .
+# git commit -m "Testing hooks"
+# git push
+
 
 # Class to handle merge requests
 class ReceiveMerge(Resource):
@@ -18,34 +22,41 @@ class ReceiveMerge(Resource):
         data = request.get_json()
         files = {}
 
-        if data["ref"] == "refs/heads/master":
-            files = get_files(data["commits"])
-            files_to_change = files[0]
-            files_removed = files[1]
-            files_to_delete = {}
+        # if data["ref"] == "refs/heads/master":
+        files = get_files(data["commits"])
+        files_to_change = files[0]
+        files_removed = files[1]
+        files_to_delete = {}
 
-            for filename in files_to_change.keys():
-                if filename in files_removed:
-                    files_to_delete[filename] = files_to_change[filename]
+        for filename in files_to_change.keys():
+            if filename in files_removed:
+                files_to_delete[filename] = files_to_change[filename]
 
-            for filename in files_to_delete.keys():
-                files_to_change.pop(filename)
+        for filename in files_to_delete.keys():
+            files_to_change.pop(filename)
 
-            print(files_to_delete)
-            print(files_to_change)
+        print(files_to_delete)
+        print(files_to_change)
 
-            for file in files_to_change.values():
-                if "Module" in file.filename and "README.md" in file.filename:
-                    parse_module(file)
+        for file in files_to_change.values():
+            if "Module" in file.filename and "Activity" not in file.filename and "README.md" in file.filename:
+                parse_module(file)
 
-            for file in files_to_delete.values():
-                if "Module" in file.filename and "README.md" in file.filename:
-                    data = md_to_json(files_to_delete[file.filename].raw_url)
-                    data["github_id"] = int(data["github_id"])
-                    requests.delete(API + "/modules", json=data)
+            if "Module" in file.filename and "Activity" in file.filename and "README.md" in file.filename:
+                parse_activity(file)
 
-            if "tests.json" in files_to_change:
-                edit_test_json(files)
+        for file in files_to_delete.values():
+            if "Module" in file.filename and "Activity" not in file.filename and "README.md" in file.filename:
+                data = md_to_json(files_to_delete[file.filename].raw_url)
+                data["github_id"] = int(data["github_id"])
+                requests.delete(API + "/modules", json=data)
+
+            if "Module" in file.filename and "Activity" in file.filename and "README.md" in file.filename:
+                data = md_to_json(files_to_delete[file.filename].raw_url)
+                data["github_id"] = int(data["github_id"])
+                requests.delete(API + "/activities", json=data)
+        # if "tests.json" in files_to_change:
+        #     edit_test_json(files)
 
         return "ok", 200
